requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
docopt==0.6.2
finta==1.3
idna==3.3
kafka-python
numpy==1.21.5
pandas==1.3.5
pipreqs==0.4.11
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
ta==0.8.0
urllib3==1.26.7
yarg==0.1.9
